requirements.txt
requests==2.32.3
pandas==2.2.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-cloud-bigquery==3.25.0
pyarrow==17.0.0
